+++ engine/sitematch/noop.go
@@ -1,0 +1,8 @@
+package sitematch
+
+type NoopSiteMatcher struct{}
+
+func (NoopSiteMatcher) Match(req SiteMatchRequest) (SiteMatchResult, error) {
+	_ = req
+	return SiteMatchResult{}, nil
+}